+++ video2dataset/main.py
@@ -17,8 +17,7 @@ from .data_writer import (
 )
 from .input_sharder import InputSharder
 from .output_sharder import OutputSharder
-from .distributor import multiprocessing_distributor, pyspark_distributor
-from .slurm_executor import SlurmDistributor
+from .distributor import multiprocessing_distributor, pyspark_distributor, SlurmDistributor
 from .workers import DownloadWorker, SubsetWorker, OpticalFlowWorker
 
 
@@ -26,8 +25,9 @@ def identity(x):
     return x
 
 
-
 # pylint: disable=unused-argument
+# pylint: disable=eval-used
+# pylint: disable=broad-except
 def video2dataset(
     url_list: str,
     output_folder: str = "videos",
@@ -65,19 +65,19 @@ def video2dataset(
     optical_flow_fps: int = -1,
     optical_flow_downsample_size: int = None,
     optical_flow_dtype: str = "np.float16",
-    sampler = None,
-    slurm_cpus_per_task: int=1,
-    slurm_job_name:str='video2dataset',
-    slurm_partition:str=None,
-    slurm_n_nodes:int=1,
-    slurm_gpus_per_node:int=8,
-    slurm_account:str=None,
-    slurm_tasks_per_node:int=1,
-    slurm_nodelist:str=None,
-    slurm_exclude:str=None,
-    slurm_cache_path:str=None,
-    slurm_timeout:int=None,
-    slurm_verbose_wait:bool=False,
+    sampler=None,
+    slurm_cpus_per_task: int = 1,
+    slurm_job_name: str = "video2dataset",
+    slurm_partition: str = None,
+    slurm_n_nodes: int = 1,
+    slurm_gpus_per_node: int = 8,
+    slurm_account: str = None,
+    slurm_tasks_per_node: int = 1,
+    slurm_nodelist: str = None,
+    slurm_exclude: str = None,
+    slurm_cache_path: str = None,
+    slurm_timeout: int = None,
+    slurm_verbose_wait: bool = False,
 ):
     """
     create video dataset from video links
@@ -86,10 +86,13 @@ def video2dataset(
     try:
         optical_flow_dtype = eval(optical_flow_dtype)
     except Exception as e:
-        print(f'Invalid optical_flow_dtype specified: {optical_flow_dtype}. Please use valid one')
-        exit(1)
+        print(f"Invalid optical_flow_dtype specified: {optical_flow_dtype}. Please use valid one")
+        print(f"{e.__class__.__name__}: {e}")
+        sys.exit(1)
 
-    assert isinstance(optical_flow_dtype, type), f'Invalid optical_flow_dtype specified: {optical_flow_dtype}. Please use valid one.'
+    assert isinstance(
+        optical_flow_dtype, type
+    ), f"Invalid optical_flow_dtype specified: {optical_flow_dtype}. Please use valid one."
 
     if sampler is None:
         sampler = identity
@@ -176,7 +179,7 @@ def video2dataset(
             number_sample_per_shard,
             done_shards,
             tmp_path,
-            sampler
+            sampler,
         )
         worker = DownloadWorker(
             sample_writer_class=sample_writer_class,
@@ -201,12 +204,7 @@ def video2dataset(
             cuts_are_clips=cuts_are_clips,
         )
     elif stage == "subset":
-        shard_iterator = OutputSharder(  # type: ignore
-            url_list,
-            input_format,
-            done_shards,
-            sampler=sampler
-        )
+        shard_iterator = OutputSharder(url_list, input_format, done_shards, sampler=sampler)  # type: ignore
         worker = SubsetWorker(  # type: ignore
             sample_writer_class=sample_writer_class,
             output_folder=output_folder,
@@ -216,12 +214,7 @@ def video2dataset(
             encode_formats=encode_formats,
         )
     elif stage == "optical_flow":
-        shard_iterator = OutputSharder(  # type: ignore
-            url_list,
-            input_format,
-            done_shards,
-            sampler=sampler
-        )
+        shard_iterator = OutputSharder(url_list, input_format, done_shards, sampler=sampler)  # type: ignore
         worker = OpticalFlowWorker(  # type: ignore
             sample_writer_class=sample_writer_class,
             output_folder=output_folder,
@@ -241,13 +234,13 @@ def video2dataset(
     called_from_slurm = False
     if distributor == "multiprocessing":
         distributor_fn = multiprocessing_distributor
-        called_from_slurm = 'GLOBAL_RANK' in os.environ
+        called_from_slurm = "GLOBAL_RANK" in os.environ
     elif distributor == "pyspark":
         distributor_fn = pyspark_distributor
     elif distributor == "slurm":
-        slurm_args = {'_'.join(key.split('_')[1:]): local_args[key] for key in local_args if key.startswith('slurm')}
-        worker_args = {key: local_args[key] for key in local_args if not key.startswith('slurm')}
-        distributor_fn = SlurmDistributor(worker_args = worker_args, **slurm_args)
+        slurm_args = {"_".join(key.split("_")[1:]): local_args[key] for key in local_args if key.startswith("slurm")}
+        worker_args = {key: local_args[key] for key in local_args if not key.startswith("slurm")}
+        distributor_fn = SlurmDistributor(worker_args=worker_args, **slurm_args)
     else:
         raise ValueError(f"Distributor {distributor} not supported")
