+++ video2dataset/output_sharder.py
@@ -14,13 +14,7 @@ class OutputSharder:
     - group_shards: the number of shards to group together
     """
 
-    def __init__(
-        self,
-        shard_list,
-        input_format,
-        done_shards,
-        sampler=lambda x: x
-    ) -> None:
+    def __init__(self, shard_list, input_format, done_shards, sampler=lambda x: x) -> None:
 
         self.input_format = input_format
         self.done_shards = done_shards
@@ -31,9 +25,9 @@ class OutputSharder:
         elif self.input_format == "files":
             self.shard_ids = [s.split("/")[-1] for s in self.shard_list]
 
-        self.shards = sampler([
-            (s_id, s) for s_id, s in zip(self.shard_ids, self.shard_list) if int(s_id) not in self.done_shards
-        ])
+        self.shards = sampler(
+            [(s_id, s) for s_id, s in zip(self.shard_ids, self.shard_list) if int(s_id) not in self.done_shards]
+        )
 
     def __iter__(self):
         """
